Incorrect ITT counter names fixed
