+++ ngraph/core/src/pass/manager.cpp
@@ -20,6 +20,7 @@
 #include "ngraph/pass/pass.hpp"
 #include "ngraph/pass/visualize_tree.hpp"
 #include "ngraph/util.hpp"
+#include "perf_counters.hpp"
 
 using namespace std;
 using namespace ngraph;
@@ -30,32 +31,6 @@ namespace ngraph
     {
         namespace
         {
-            class PerfCounters
-            {
-                PerfCounters(PerfCounters const&) = delete;
-                PerfCounters& operator=(PerfCounters const&) = delete;
-
-            public:
-                PerfCounters() = default;
-
-                openvino::itt::handle_t operator[](::ngraph::Node::type_info_t const& type_inf)
-                {
-                    std::lock_guard<std::mutex> guard(m_mutex);
-                    auto it = m_counters.find(&type_inf);
-                    if (it != m_counters.end())
-                        return it->second;
-                    return m_counters[&type_inf] = openvino::itt::handle(type_inf.name);
-                }
-
-            private:
-                using key = ::ngraph::Node::type_info_t const*;
-                using value = openvino::itt::handle_t;
-                using counters_map = std::unordered_map<key, value>;
-
-                std::mutex m_mutex;
-                counters_map m_counters;
-            };
-
             PerfCounters& perf_counters()
             {
                 static PerfCounters counters;
